+++ core/src/commonMain/kotlin/org/micoli/micraft/world/BlockDefinition.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class BlockDefinition(
-    val hardness: Int = 1,
+    val hardness: Float = 1f,
     val solid: Boolean = true,
     val transparent: Boolean = false,
     val minimapColor: List<Int> = listOf(128, 128, 128),
